refactor: DI部分の分割と、修正
